+++ CHANGELOG.md
@@ -10,7 +10,9 @@
 
 - `pipe(lazyPromise, a, b)` becomes a slightly more ergonomic `lazyPromise.pipe(a, b)`.
 
-- `lazy` and `eager` renamed to `fromEager` and `toEager`.
+- `lazy` and `eager` are renamed to `fromEager` and `toEager` and now use an options object.
+
+- `fromEager` flattens the returned lazy promise.
 
 ## [@lazy-promise/core@0.0.16](https://github.com/lazy-promise/lazy-promise/tree/%40lazy-promise/core%400.0.16)